Restructures tests and added a few
